refactor: MemberGoal 도메인 분리
